+++ modules/contracts/src.ts/artifacts.ts
@@ -1,5 +1,5 @@
 import { HexString } from "@connext/vector-types";
-import { utils } from "ethers";
+import { FunctionFragment, EventFragment, ParamType } from "@ethersproject/abi";
 
 import * as AssetTransfer from "../artifacts/src.sol/AssetTransfer.sol/AssetTransfer.json";
 import * as ChannelFactory from "../artifacts/src.sol/ChannelFactory.sol/ChannelFactory.json";
@@ -16,7 +16,7 @@ import * as VectorChannel from "../artifacts/src.sol/interfaces/IVectorChannel.s
 import * as Withdraw from "../artifacts/src.sol/transferDefinitions/Withdraw.sol/Withdraw.json";
 import * as TestLibIterableMapping from "../artifacts/src.sol/testing/TestLibIterableMapping.sol/TestLibIterableMapping.json";
 
-type Abi = Array<string | utils.FunctionFragment | utils.EventFragment | utils.ParamType>;
+type Abi = Array<string | FunctionFragment | EventFragment | ParamType>;
 
 type Artifact = {
   contractName: string;