+++ modules/contracts/src.ts/tests/cmcs/withdraw.spec.ts
@@ -1,11 +1,14 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
-import { ChannelSigner, expect, getRandomBytes32 } from "@connext/vector-utils";
+import { ChannelSigner, expect, getRandomAddress, getRandomBytes32 } from "@connext/vector-utils";
+import { BigNumber } from "@ethersproject/bignumber";
 import { AddressZero } from "@ethersproject/constants";
-import { BigNumber, Contract, Wallet } from "ethers";
-import { parseEther } from "ethers/lib/utils";
+import { Contract } from "@ethersproject/contracts";
+import { parseEther } from "@ethersproject/units";
+import { Wallet } from "@ethersproject/wallet";
 
 import { getTestAddressBook, getTestChannel, alice } from "..";
 import { deployContracts, WithdrawCommitment } from "../..";
+import { AddressBook } from "../../addressBook";
 import { bob, provider } from "../constants";
 
 describe("CMCWithdraw.sol", function() {
@@ -14,9 +17,10 @@ describe("CMCWithdraw.sol", function() {
 
   let channel: Contract;
   let failingToken: Contract;
+  let addressBook: AddressBook;
 
   beforeEach(async () => {
-    const addressBook = await getTestAddressBook();
+    addressBook = await getTestAddressBook();
     channel = await getTestChannel(addressBook);
 
     await deployContracts(alice, addressBook, [["FailingToken", []]]);
@@ -29,9 +33,17 @@ describe("CMCWithdraw.sol", function() {
 
     const tokenTx = await failingToken.mint(channel.address, parseEther("0.001"));
     await tokenTx.wait();
+
+    // Make transfers pass
+    const dontRevert = await failingToken.setTransferShouldRevert(false);
+    await dontRevert.wait();
+    const dontFail = await failingToken.setTransferShouldFail(false);
+    await dontFail.wait();
+    const dontRejectEther = await failingToken.setRejectEther(false);
+    await dontRejectEther.wait();
   });
 
-  it("should work", async () => {
+  it("should work for Ether", async () => {
     const preWithdrawRecipient = await provider.getBalance(recipient);
     const preWithdrawChannel = await provider.getBalance(channel.address);
     const withdrawAmount = BigNumber.from(1000);
@@ -48,19 +60,111 @@ describe("CMCWithdraw.sol", function() {
 
     const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(commitment.hashToSign());
     const bobSig = await new ChannelSigner(bob.privateKey).signMessage(commitment.hashToSign());
-    expect(await channel.getWithdrawalTransactionRecord(recipient, AddressZero, withdrawAmount, nonce)).to.be.false;
 
-    const tx = await channel.withdraw(recipient, AddressZero, withdrawAmount, nonce, aliceSig, bobSig);
+    const withdrawData = commitment.getWithdrawData();
+    expect(await channel.getWithdrawalTransactionRecord(withdrawData)).to.be.false;
+
+    const tx = await channel.withdraw(withdrawData, aliceSig, bobSig);
     await tx.wait();
 
     expect(await provider.getBalance(recipient)).to.be.eq(preWithdrawRecipient.add(withdrawAmount));
     expect(await provider.getBalance(channel.address)).to.be.eq(preWithdrawChannel.sub(withdrawAmount));
     expect(await channel.getTotalTransferred(AddressZero)).to.be.eq(withdrawAmount);
-    expect(await channel.getWithdrawalTransactionRecord(recipient, AddressZero, withdrawAmount, nonce)).to.be.true;
+    expect(await channel.getWithdrawalTransactionRecord(withdrawData)).to.be.true;
   });
 
-  it("should fail if the tx has already been executed", async () => {
+  it("should work for standard-compliant tokens", async () => {
+    const preWithdrawRecipient = await failingToken.balanceOf(recipient);
+    const preWithdrawChannel = await failingToken.balanceOf(channel.address);
     const withdrawAmount = BigNumber.from(1000);
+    const nonce = BigNumber.from(1);
+    const commitment = new WithdrawCommitment(
+      channel.address,
+      alice.address,
+      bob.address,
+      recipient,
+      failingToken.address,
+      withdrawAmount.toString(),
+      nonce.toString(),
+    );
+
+    const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(commitment.hashToSign());
+    const bobSig = await new ChannelSigner(bob.privateKey).signMessage(commitment.hashToSign());
+
+    const withdrawData = commitment.getWithdrawData();
+    expect(await channel.getWithdrawalTransactionRecord(withdrawData)).to.be.false;
+
+    const tx = await channel.withdraw(withdrawData, aliceSig, bobSig);
+    await tx.wait();
+
+    expect(await failingToken.balanceOf(recipient)).to.be.eq(preWithdrawRecipient.add(withdrawAmount));
+    expect(await failingToken.balanceOf(channel.address)).to.be.eq(preWithdrawChannel.sub(withdrawAmount));
+    expect(await channel.getTotalTransferred(failingToken.address)).to.be.eq(withdrawAmount);
+    expect(await channel.getWithdrawalTransactionRecord(withdrawData)).to.be.true;
+  });
+
+  it("should work for missing-return-value-bug tokens", async () => {
+    await deployContracts(alice, addressBook, [["NonconformingToken", []]]);
+    const nonconformingToken = addressBook.getContract("NonconformingToken");
+    await nonconformingToken.mint(alice.address, parseEther("0.001"));
+
+    // Send tokens to channel
+    const tokenTx = await nonconformingToken.mint(channel.address, parseEther("0.001"));
+    await tokenTx.wait();
+
+    const preWithdrawRecipient = await nonconformingToken.balanceOf(recipient);
+    const preWithdrawChannel = await nonconformingToken.balanceOf(channel.address);
+    const withdrawAmount = BigNumber.from(1000);
+    const nonce = BigNumber.from(1);
+    const commitment = new WithdrawCommitment(
+      channel.address,
+      alice.address,
+      bob.address,
+      recipient,
+      nonconformingToken.address,
+      withdrawAmount.toString(),
+      nonce.toString(),
+    );
+
+    const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(commitment.hashToSign());
+    const bobSig = await new ChannelSigner(bob.privateKey).signMessage(commitment.hashToSign());
+
+    const withdrawData = commitment.getWithdrawData();
+    expect(await channel.getWithdrawalTransactionRecord(withdrawData)).to.be.false;
+
+    const tx = await channel.withdraw(withdrawData, aliceSig, bobSig);
+    await tx.wait();
+
+    expect(await nonconformingToken.balanceOf(recipient)).to.be.eq(preWithdrawRecipient.add(withdrawAmount));
+    expect(await nonconformingToken.balanceOf(channel.address)).to.be.eq(preWithdrawChannel.sub(withdrawAmount));
+    expect(await channel.getTotalTransferred(nonconformingToken.address)).to.be.eq(withdrawAmount);
+    expect(await channel.getWithdrawalTransactionRecord(withdrawData)).to.be.true;
+  });
+
+  it("should fail for wrong channel address", async () => {
+    const withdrawAmount = BigNumber.from(1000);
+    const nonce = BigNumber.from(1);
+    const commitment = new WithdrawCommitment(
+      getRandomAddress(),
+      alice.address,
+      bob.address,
+      recipient,
+      AddressZero,
+      withdrawAmount.toString(),
+      nonce.toString(),
+    );
+
+    const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(commitment.hashToSign());
+    const bobSig = await new ChannelSigner(bob.privateKey).signMessage(commitment.hashToSign());
+
+    const withdrawData = commitment.getWithdrawData();
+    await expect(channel.withdraw(withdrawData, aliceSig, bobSig)).revertedWith(
+      "CMCWithdraw: Channel address mismatch",
+    );
+  });
+
+  it("should fail if it is a no-op (callTo == address(0) && amount == 0)", async () => {
+    const withdrawAmount = BigNumber.from(0);
     const nonce = BigNumber.from(1);
     const commitment = new WithdrawCommitment(
       channel.address,
@@ -75,10 +179,8 @@ describe("CMCWithdraw.sol", function() {
     const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(commitment.hashToSign());
     const bobSig = await new ChannelSigner(bob.privateKey).signMessage(commitment.hashToSign());
 
-    await channel.withdraw(recipient, AddressZero, withdrawAmount, nonce, aliceSig, bobSig);
-    await expect(channel.withdraw(recipient, AddressZero, withdrawAmount, nonce, aliceSig, bobSig)).revertedWith(
-      "CMCWithdraw: Transaction has already been executed",
-    );
+    const withdrawData = commitment.getWithdrawData();
+    await expect(channel.withdraw(withdrawData, aliceSig, bobSig)).revertedWith("CMCWithdraw: No-op");
   });
 
   it("should fail if alice signature is invalid", async () => {
@@ -97,9 +199,8 @@ describe("CMCWithdraw.sol", function() {
     const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(getRandomBytes32());
     const bobSig = await new ChannelSigner(bob.privateKey).signMessage(commitment.hashToSign());
 
-    await expect(channel.withdraw(recipient, AddressZero, withdrawAmount, nonce, aliceSig, bobSig)).revertedWith(
-      "CMCWithdraw: Invalid alice signature",
-    );
+    const withdrawData = commitment.getWithdrawData();
+    await expect(channel.withdraw(withdrawData, aliceSig, bobSig)).revertedWith("CMCWithdraw: Invalid alice signature");
   });
 
   it("should fail if bob signature is invalid", async () => {
@@ -118,12 +219,64 @@ describe("CMCWithdraw.sol", function() {
     const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(commitment.hashToSign());
     const bobSig = await new ChannelSigner(bob.privateKey).signMessage(getRandomBytes32());
 
-    await expect(channel.withdraw(recipient, AddressZero, withdrawAmount, nonce, aliceSig, bobSig)).revertedWith(
-      "CMCWithdraw: Invalid bob signature",
+    const withdrawData = commitment.getWithdrawData();
+    await expect(channel.withdraw(withdrawData, aliceSig, bobSig)).revertedWith("CMCWithdraw: Invalid bob signature");
+  });
+
+  it("should fail if the tx has already been executed", async () => {
+    const withdrawAmount = BigNumber.from(1000);
+    const nonce = BigNumber.from(1);
+    const commitment = new WithdrawCommitment(
+      channel.address,
+      alice.address,
+      bob.address,
+      recipient,
+      AddressZero,
+      withdrawAmount.toString(),
+      nonce.toString(),
+    );
+
+    const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(commitment.hashToSign());
+    const bobSig = await new ChannelSigner(bob.privateKey).signMessage(commitment.hashToSign());
+
+    const withdrawData = commitment.getWithdrawData();
+    await channel.withdraw(withdrawData, aliceSig, bobSig);
+    await expect(channel.withdraw(withdrawData, aliceSig, bobSig)).revertedWith(
+      "CMCWithdraw: Transaction has already been executed",
     );
   });
 
-  it("should fail if transfer fails", async () => {
+  it("should fail if Ether transfer reverts", async () => {
+    const rejectEtherRecipient = failingToken.address;
+
+    // Make transfers fail
+    const rejectEther = await failingToken.setRejectEther(true);
+    await rejectEther.wait();
+
+    const withdrawAmount = BigNumber.from(1000);
+    const nonce = BigNumber.from(1);
+    const commitment = new WithdrawCommitment(
+      channel.address,
+      alice.address,
+      bob.address,
+      rejectEtherRecipient,
+      AddressZero,
+      withdrawAmount.toString(),
+      nonce.toString(),
+    );
+
+    const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(commitment.hashToSign());
+    const bobSig = await new ChannelSigner(bob.privateKey).signMessage(commitment.hashToSign());
+    const withdrawData = commitment.getWithdrawData();
+
+    await expect(channel.withdraw(withdrawData, aliceSig, bobSig)).revertedWith("Ether rejected");
+  });
+
+  it("should fail if token transfer fails", async () => {
+    // Make transfers fail
+    const failing = await failingToken.setTransferShouldFail(true);
+    await failing.wait();
+
     const withdrawAmount = BigNumber.from(1000);
     const nonce = BigNumber.from(1);
     const commitment = new WithdrawCommitment(
@@ -138,9 +291,34 @@ describe("CMCWithdraw.sol", function() {
 
     const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(commitment.hashToSign());
     const bobSig = await new ChannelSigner(bob.privateKey).signMessage(commitment.hashToSign());
+    const withdrawData = commitment.getWithdrawData();
 
-    await expect(
-      channel.withdraw(recipient, failingToken.address, withdrawAmount, nonce, aliceSig, bobSig),
-    ).revertedWith("FAIL: Failing token");
+    await expect(channel.withdraw(withdrawData, aliceSig, bobSig)).revertedWith("CMCWithdraw: Transfer failed");
   });
+
+  it("should fail if token transfer reverts", async () => {
+    // Make transfers revert
+    const reverting = await failingToken.setTransferShouldRevert(true);
+    await reverting.wait();
+
+    const withdrawAmount = BigNumber.from(1000);
+    const nonce = BigNumber.from(1);
+    const commitment = new WithdrawCommitment(
+      channel.address,
+      alice.address,
+      bob.address,
+      recipient,
+      failingToken.address,
+      withdrawAmount.toString(),
+      nonce.toString(),
+    );
+
+    const aliceSig = await new ChannelSigner(alice.privateKey).signMessage(commitment.hashToSign());
+    const bobSig = await new ChannelSigner(bob.privateKey).signMessage(commitment.hashToSign());
+    const withdrawData = commitment.getWithdrawData();
+
+    await expect(channel.withdraw(withdrawData, aliceSig, bobSig)).revertedWith("FAIL: Failing token");
+  });
+
+  it.skip("should call helper contract if given non-zero address", async () => {});
 });