+++ modules/types/src/network.ts
@@ -1,9 +1,9 @@
-import { providers } from "ethers";
+import { JsonRpcProvider } from "@ethersproject/providers";
 
 export type ChainProviders = {
   [chainId: number]: string;
 };
 
 export type HydratedProviders = {
-  [chainId: number]: providers.JsonRpcProvider;
+  [chainId: number]: JsonRpcProvider;
 };